Tests: Setup (#23)
### Features
- Add testcontainers dependencies;
- Add `image` field to docker compose;
- Add `spring-security-test` as dependency;
- Add `IntegrationTestConfig`, `UnitTestConfig` and `RepositoryUnitTestConfig` for test setup;
- Add `DatabaseContainerInitializer` to fake the database;
- Add password and role validation to `/auth/authenticate`;
- Add application-test.properties;
- Add getter and setter to `registrationIsEnable`;

### Fixes
- Fix DELETE not checking if item exists;
- Fix wrong `/auth/register` status code;
- Fix wrong save status code in all controllers;

### Changes
- Delete HawAPIApplicationTests;
- Replace byte to Byte in some models;
- Update MappersConfig (Add more configurations);

+++ src/main/java/com/lucasjosino/hawapi/models/EpisodeModel.java
@@ -19,11 +19,11 @@ public class EpisodeModel extends BaseModel {
     private String description;
 
     @Column(nullable = false)
-    private int duration;
+    private Integer duration;
 
     @JsonProperty("episode_num")
     @Column(nullable = false, name = "episode_num")
-    private int episodeNum;
+    private Byte episodeNum;
 
     @JsonProperty("next_episode")
     @Column(name = "next_episode")
@@ -59,19 +59,19 @@ public class EpisodeModel extends BaseModel {
         this.description = description;
     }
 
-    public int getDuration() {
+    public Integer getDuration() {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(Integer duration) {
         this.duration = duration;
     }
 
-    public int getEpisodeNum() {
+    public Byte getEpisodeNum() {
         return episodeNum;
     }
 
-    public void setEpisodeNum(int episodeNum) {
+    public void setEpisodeNum(Byte episodeNum) {
         this.episodeNum = episodeNum;
     }
